# docs/swagger.yaml
basePath: /api
definitions:
...
  repository.AccountType:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - AccountTypeLocal
    - AccountTypeService
    - AccountTypeOIDC
...
  userhandler.APIKeyCreatedResponse:
    properties:
      created_at:
        type: integer
      expires_at:
        description: nullable
        type: integer
      id:
        type: string
      key_hint:
        type: string
      last_used_at:
        description: nullable
        type: integer
      name:
        type: string
      scope_admin:
        type: boolean
      scope_create:
        type: boolean
      scope_delete:
        type: boolean
      scope_edit:
        type: boolean
      scope_view:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/userhandler.UserSubResponse'
    type: object
  userhandler.APIKeyResponse:
    properties:
      created_at:
        type: integer
      expires_at:
        description: nullable
        type: integer
      id:
        type: string
      key_hint:
        type: string
      last_used_at:
        description: nullable
        type: integer
      name:
        type: string
      scope_admin:
        type: boolean
      scope_create:
        type: boolean
      scope_delete:
        type: boolean
      scope_edit:
        type: boolean
      scope_view:
        type: boolean
      user:
        $ref: '#/definitions/userhandler.UserSubResponse'
    type: object
  userhandler.CreateAPIKeyPayload:
    properties:
      expires_at:
        type: integer
      name:
        type: string
      scope_admin:
        type: boolean
      scope_create:
        type: boolean
      scope_delete:
        type: boolean
      scope_edit:
        type: boolean
      scope_view:
        type: boolean
    type: object
  userhandler.CreateUserPayload:
    properties:
      account_type:
        $ref: '#/definitions/repository.AccountType'
      is_admin:
        type: boolean
      password:
        type: string
      permissions:
        items:
          $ref: '#/definitions/userhandler.DatabasePermission'
        type: array
      username:
        type: string
    type: object
  userhandler.DatabasePermission:
    properties:
      can_admin:
        type: boolean
      can_create:
        type: boolean
      can_delete:
        type: boolean
      can_edit:
        type: boolean
      can_view:
        type: boolean
      database_id:
        type: string
    type: object
  userhandler.UpdateAPIKeyPayload:
    properties:
      expires_at:
        type: integer
      name:
        type: string
      scope_admin:
        type: boolean
      scope_create:
        type: boolean
      scope_delete:
        type: boolean
      scope_edit:
        type: boolean
      scope_view:
        type: boolean
    type: object
...
  userhandler.UpdateUserPayload:
    properties:
      is_admin:
        type: boolean
      password:
        type: string
      permissions:
        items:
          $ref: '#/definitions/userhandler.DatabasePermission'
        type: array
      username:
        type: string
    type: object
  userhandler.UserResponse:
    properties:
      account_type:
        $ref: '#/definitions/repository.AccountType'
      id:
        type: string
      is_admin:
        type: boolean
      permissions:
        items:
          $ref: '#/definitions/userhandler.DatabasePermission'
        type: array
      username:
        type: string
    type: object
  userhandler.UserSubResponse:
    properties:
      account_type:
        $ref: '#/definitions/repository.AccountType'
      id:
        type: string
      is_admin:
        type: boolean
      username:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: denglerchr@gmail.com
    name: Christian Dengler
    url: https://www.swcd.lu
  description: This is a server for a image, audio and file storage with integrated
    web-ui.
  title: SWCD MediaHub-API
  version: 3.1.0
paths:
...
  /user:
    delete:
      description: Deletes a user account and cascades the deletion to remove their
        database permissions. Requires the global IsAdmin role.
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Missing or invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Cannot delete the last remaining admin user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete a user
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Creates a new user account and optionally defines their database
        permissions. Requires the global IsAdmin role.
      parameters:
      - description: New user details and permissions
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: User successfully created
          schema:
            $ref: '#/definitions/userhandler.UserResponse'
        "400":
          description: Invalid JSON body, missing fields, or duplicate database names
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User
  /user/{user_ulid}:
    get:
      description: Retrieves a specific user account and their database permissions.
        Requires the global IsAdmin role.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User record
          schema:
            $ref: '#/definitions/userhandler.UserResponse'
        "400":
          description: Missing or invalid user ULID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Retrieve a user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Updates an existing user's global status, password, or database
        permissions. Permissions act as an Upsert/Replace operation. Requires the
        global IsAdmin role.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.UpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully updated
          schema:
            $ref: '#/definitions/userhandler.UserResponse'
        "400":
          description: Missing ID, duplicate DBs, or invalid JSON body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Cannot remove last admin user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update an existing user
      tags:
      - User
  /user/{user_ulid}/keys:
    get:
      description: Retrieves the complete list of keys associated with the specified
        user. Hint only, actual key is hashed. Requires admin or self ownership.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userhandler.APIKeyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve API keys for a user
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Generates a new API key for the specified user and returns the
        plaintext token. Plentext token is returned only once. Requires admin or self
        ownership.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      - description: API Key details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.CreateAPIKeyPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/userhandler.APIKeyCreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new API key
      tags:
      - User
  /user/{user_ulid}/keys/{key_ulid}:
    delete:
      description: Deletes/revokes an existing API key. Requires admin or self ownership.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      - description: Key ULID
        in: path
        name: key_ulid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete/revoke an API key
      tags:
      - User
    get:
      description: Retrieves the metadata of a specific API key. Requires admin or
        self ownership.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      - description: Key ULID
        in: path
        name: key_ulid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.APIKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve a specific API key details
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Updates an existing API key's name, expiry timestamp, or scopes.
        Requires admin or self ownership.
      parameters:
      - description: User ULID
        in: path
        name: user_ulid
        required: true
        type: string
      - description: Key ULID
        in: path
        name: key_ulid
        required: true
        type: string
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.UpdateAPIKeyPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.APIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an API key
      tags:
      - User
...
  /users:
    get:
      description: Retrieves a list of all user accounts and their assigned database
        permissions. Requires the global IsAdmin role.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/userhandler.UserResponse'
            type: array
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: User lacks IsAdmin role'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Retrieve all users
      tags:
      - User
  /users/keys:
    get:
      description: Retrieves all active API keys in the system with associated user
        details. Requires global admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userhandler.APIKeyResponse'
            type: array
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve all API keys
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"